Isolate node controller from GUI implementation
